Rework how descriptors work in D3D12
We used to generate them on the fly bug this is a big chunk of memory to store the descriptor
Instead, preallocate them where we can for textures, samplers, RW textures, etc
Buffers don't preallocate due to the offset but are fast to generate views from and don't need much data
Still pending is to separate the CPU-GPU descriptor coupling which in hindsight was a mistake

+++ Source/Rendering/d3d12/CrD3D12.h
@@ -75,9 +75,10 @@ namespace crd3d
 
 	// A shader-visible heap will have two handles, CPU and GPU. The CPU handle is what we use to update
 	// the data in the descriptor, the GPU handle is what we use to bind it to the command buffer
+
 	struct DescriptorD3D12
 	{
-		DescriptorD3D12 operator + (uint32_t offset) const
+		DescriptorD3D12 operator + (size_t offset) const
 		{
 			DescriptorD3D12 descriptor = *this;
 			descriptor.cpuHandle.ptr += offset;
@@ -85,7 +86,7 @@ namespace crd3d
 			return descriptor;
 		}
 
-		DescriptorD3D12& operator += (uint32_t offset)
+		DescriptorD3D12& operator += (size_t offset)
 		{
 			*this = *this + offset;
 			return *this;
